Move some wasm code to it's own module
